+++ Mobile/package.json
@@ -8,29 +8,29 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@react-native-async-storage/async-storage": "~1.15.0",
+    "@react-native-async-storage/async-storage": "~1.17.3",
     "@react-native-community/masked-view": "0.1.10",
     "@react-native-firebase/app": "^12.8.0",
     "@react-native-firebase/storage": "^12.8.0",
     "@react-navigation/stack": "^6.0.7",
-    "expo": "~42.0.1",
-    "expo-app-loading": "^1.1.2",
-    "expo-av": "~9.2.3",
-    "expo-font": "~9.2.1",
-    "expo-status-bar": "~1.0.4",
+    "expo": "^47.0.0",
+    "expo-app-loading": "~2.1.1",
+    "expo-av": "~13.0.1",
+    "expo-font": "~11.0.1",
+    "expo-status-bar": "~1.4.2",
     "firebase": "8.2.3",
-    "react": "16.13.1",
-    "react-dom": "^16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-native": "0.70.5",
     "react-native-audio-recorder-player": "^3.2.0",
-    "react-native-gesture-handler": "~1.10.2",
+    "react-native-gesture-handler": "~2.8.0",
     "react-native-ionicons": "^4.6.5",
     "react-native-paper": "^4.9.2",
-    "react-native-reanimated": "~2.2.0",
-    "react-native-safe-area-context": "3.2.0",
-    "react-native-screens": "~3.4.0",
+    "react-native-reanimated": "~2.12.0",
+    "react-native-safe-area-context": "4.4.1",
+    "react-native-screens": "~3.18.0",
     "react-native-vector-icons": "^8.1.0",
-    "react-native-web": "^0.13.18",
+    "react-native-web": "~0.18.7",
     "react-navigation": "^4.4.4",
     "react-navigation-stack": "^2.10.4",
     "react-redux": "^7.2.5",
@@ -38,7 +38,7 @@
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
-    "@babel/core": "~7.9.0"
+    "@babel/core": "^7.19.3"
   },
   "private": true
 }
